+++ metadrive/component/vehicle_module/mini_map.py
@@ -7,7 +7,7 @@ from metadrive.engine.engine_utils import get_global_config, engine_initialized
 
 class MiniMap(BaseCamera):
     CAM_MASK = CamMask.MiniMap
-    display_region_size = [0., 1 / 3, BaseCamera.display_bottom, BaseCamera.display_top]
+    display_region_size = [0., 1 / 3, 0.8, 1.0]
 
     def __init__(self):
         assert engine_initialized(), "You should initialize engine before adding camera to vehicle"
